# assets/openapi.yaml
openapi: 3.1.0
info:
  title: openccu-loom REST API
  version: 1.0.0
  description: |
    North-bound REST + WebSocket surface of the openccu-loom daemon.
    Errors follow RFC 9457 problem+json.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: /api/v1
security:
  - basicAuth: []
  - bearerAuth: []
  - openIdConnect: [viewer]
...
paths:
...
  /users:
    get:
      operationId: listUsersV2
      summary: List all users (admin)
      tags: [user-admin]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserSummary" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: createUser
      summary: Create a new user (admin)
      tags: [user-admin]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreate" }
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserSummary" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "500": { $ref: "#/components/responses/InternalError" }
  /users/{subject}:
    parameters:
      - name: subject
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: patchUser
      summary: Update user password or role (admin)
      tags: [user-admin]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteUser
      summary: Delete a user (admin)
      description: |
        Returns 409 Conflict when the deletion would remove the last
        admin account. The daemon enforces at least one admin at all
        times.
      tags: [user-admin]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      responses:
        "204":
          description: Deleted
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot remove the last admin account
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "500": { $ref: "#/components/responses/InternalError" }

  # ---- Token management v2 (Wave-C) ----
  /auth/tokens/v2:
    get:
      operationId: listTokensV2
      summary: List API tokens with fingerprint + metadata (admin)
      tags: [token-admin]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      responses:
        "200":
          description: Token list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TokenSummary" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: createTokenV2
      summary: Issue a new bearer token (admin) — plaintext shown exactly once
      description: |
        Generates a fresh bearer token bound to the supplied subject
        and role. The raw token value is returned exactly once in the
        201 response — the daemon stores only an opaque fingerprint
        thereafter. Clients MUST capture the `token` field on creation.
      tags: [token-admin]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenCreate" }
      responses:
        "201":
          description: Token created — raw value visible exactly once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenCreated" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Username + password (operator-grade or above for mutations).
        The role hierarchy (viewer ⊂ operator ⊂ admin) is enforced
        server-side regardless of authentication scheme.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API token issued via the auth tokens surface (admin-only
        provisioning). Token's stored role determines authorisation.
    openIdConnect:
      type: openIdConnect
      description: |
        OpenID Connect authentication. The operator configures the
        OIDC discovery URL; the placeholder below is the standard
        well-known path. Per-operation `security:` blocks express
        the **minimum role** the daemon enforces via OIDC-style
        scopes — the same enforcement applies when the client
        authenticates via basicAuth or bearerAuth, but those schemes
        cannot natively express scopes in OpenAPI 3.x.

        Scope vocabulary mirrors `Identity.role`:
          - `viewer` — read-only operations
          - `operator` — mutations (write data points, manage links,
            edit schedules, run config-import, trigger refresh)
          - `admin` — dangerous operations (delete device, edit log
            levels, list users/tokens, trigger backup, manage
            interfaces, edit Matter fabrics)
      openIdConnectUrl: /.well-known/openid-configuration
...
  responses:
    BadRequest:
      description: Validation or parsing failure
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    InternalError:
      description: Unexpected server error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 9457 problem detail. The `type` URI selects one of the
        canonical problem kinds the daemon emits; the `code` member
        carries the short tag (also surfaced as the `X-Problem-Code`
        response header). Clients should switch on `code` for typed
        error handling — the `type` URI is the human-readable form.

        **Rate-limiting:** when the daemon emits `code: rate_limited`
        (HTTP 429) it sets the `Retry-After` response header per
        RFC 9110 §10.2.3 with the integer number of seconds the
        client should wait before retrying. Clients that respect
        the header are immune to the throttle; clients that don't
        will see repeated 429s with the same `Retry-After` value
        until the throttle window expires.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable URI for one of the canonical problem kinds. All
            URIs share the base `https://openccu-loom.dev/errors/`
            and the trailing segment matches the `code` enum below.
            The base URI is a stable identifier, not a fetchable
            resource — it is not guaranteed to resolve to a document.
          enum:
            - https://openccu-loom.dev/errors/validation
            - https://openccu-loom.dev/errors/not_found
            - https://openccu-loom.dev/errors/conflict
            - https://openccu-loom.dev/errors/unauthorized
            - https://openccu-loom.dev/errors/forbidden
            - https://openccu-loom.dev/errors/unsupported
            - https://openccu-loom.dev/errors/rate_limited
            - https://openccu-loom.dev/errors/internal
            - https://openccu-loom.dev/errors/bad_request
            - https://openccu-loom.dev/errors/service_unready
            - https://openccu-loom.dev/errors/upstream_unavailable
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Short tag (also surfaced as the `X-Problem-Code` response header).
          enum:
            - validation
            - not_found
            - conflict
            - unauthorized
            - forbidden
            - unsupported
            - rate_limited
            - internal
            - bad_request
            - service_unready
            - upstream_unavailable
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
              min: {}
              max: {}
...
    UserCreate:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          description: Login name; must be unique.
        password:
          type: string
          description: Plaintext password; the daemon hashes it before storage.
        role:
          type: string
          enum: [viewer, operator, admin]
    UserUpdate:
      type: object
      description: All fields optional; omitted fields leave the user unchanged.
      properties:
        password:
          type: string
          description: New plaintext password. Omit to leave unchanged.
        role:
          type: string
          enum: [viewer, operator, admin]
          description: New role. Omit to leave unchanged.
    UserSummary:
      type: object
      required: [subject, role, created_at]
      properties:
        subject:
          type: string
          description: Login name / stable identity key.
        role:
          type: string
          enum: [viewer, operator, admin]
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          description: Last successful authentication time. Absent when never authenticated.
    TokenCreate:
      type: object
      required: [subject, role]
      properties:
        subject:
          type: string
          description: Logical owner of the token (e.g. "homeassistant", "ci-runner").
        role:
          type: string
          enum: [viewer, operator, admin]
    TokenCreated:
      type: object
      required: [token, fingerprint]
      description: |
        Returned by `POST /auth/tokens/v2`. The `token` field carries the
        raw secret and is exposed exactly once on creation.
      properties:
        token:
          type: string
          description: Raw bearer token. URL-safe base64. Shown EXACTLY ONCE — store it immediately.
        fingerprint:
          type: string
          description: Stable opaque identifier derived from the token (sha256-based).
    TokenSummary:
      type: object
      required: [fingerprint, subject, role, created_at]
      properties:
        fingerprint:
          type: string
          description: Stable opaque identifier used as the delete key.
        subject:
          type: string
        role:
          type: string
          enum: [viewer, operator, admin]
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          description: Last successful authentication time. Absent when never used.
...