# internal/north/rest/handlers/admin_users.go
// SPDX-License-Identifier: MIT
// Copyright (C) 2026 OpenCCU-Loom authors.

package handlers

import (
	"context"
	"errors"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"

	"github.com/SukramJ/openccu-loom/internal/audit"
	"github.com/SukramJ/openccu-loom/internal/auth"
	"github.com/SukramJ/openccu-loom/internal/north/rest/problem"
	"github.com/SukramJ/openccu-loom/internal/store/sqlite"
)

// UserAdminService is the DI surface for the /api/v1/admin/users
// endpoints. The router wires a [*sqlite.UserStore] underneath.
type UserAdminService interface {
	// Put creates or updates a user with the given credentials and role.
	Put(ctx context.Context, subject, password string, role auth.Role) error
	// Delete removes a user. Returns [sqlite.ErrUserNotFound] when the
	// subject is unknown and [sqlite.ErrLastAdmin] when removing the
	// subject would leave zero admins.
	Delete(ctx context.Context, subject string) error
	// List returns every user sorted by subject.
	List(ctx context.Context) ([]sqlite.UserRow, error)
	// Count returns the number of users in the store.
	Count(ctx context.Context) (int, error)
}

// createUserRequest is the body of POST /admin/users.
type createUserRequest struct {
	Username string    `json:"username"`
	Password string    `json:"password"`
	Role     auth.Role `json:"role"`
}

// updateUserRequest is the body of PATCH /admin/users/{subject}.
type updateUserRequest struct {
	Password string    `json:"password"`
	Role     auth.Role `json:"role"`
}

// userSummaryResponse is the response for a successful user creation.
type userSummaryResponse struct {
	Subject   string    `json:"subject"`
	Role      auth.Role `json:"role"`
	CreatedAt time.Time `json:"created_at,omitempty"`
}

// userListEntry is one element of the GET /admin/users response.
type userListEntry struct {
	Subject    string     `json:"subject"`
	Role       auth.Role  `json:"role"`
	CreatedAt  time.Time  `json:"created_at"`
	LastSeenAt *time.Time `json:"last_seen_at,omitempty"`
}

// validRole reports whether r is one of the three canonical roles.
func validRole(r auth.Role) bool {
	switch r {
	case auth.RoleAdmin, auth.RoleOperator, auth.RoleViewer:
		return true
	}
	return false
}

// CreateUser handles POST /admin/users. It creates or upserts a user
// with the supplied credentials. Returns 201 with a summary on success.
func CreateUser(svc UserAdminService, rec audit.Recorder) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var body createUserRequest
		if err := DecodeJSON(r, &body); err != nil {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Invalid request body", err.Error()))
			return
		}
		if body.Username == "" {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Missing username", "username is required"))
			return
		}
		if body.Password == "" {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Missing password", "password is required"))
			return
		}
		if !validRole(body.Role) {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Invalid role",
					"role must be one of: admin, operator, viewer"))
			return
		}
		if err := svc.Put(r.Context(), body.Username, body.Password, body.Role); err != nil {
			problem.Write(w, http.StatusInternalServerError,
				problem.New(problem.TypeInternal, r, "User creation failed", err.Error()))
			return
		}
		actor := identityFromCtx(r.Context())
		if rec != nil {
			rec.Record(audit.Entry{
				User:   actor,
				Action: audit.ActionUserCreate,
				Note:   "subject=" + body.Username + " role=" + string(body.Role),
			})
		}
		JSON(w, http.StatusCreated, userSummaryResponse{
			Subject: body.Username,
			Role:    body.Role,
		})
	}
}

// UpdateUser handles PATCH /admin/users/{subject}. Requires both
// password and role in the request body to perform a full upsert.
// Returns 404 when the subject does not exist.
func UpdateUser(svc UserAdminService, rec audit.Recorder) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		subject := chi.URLParam(r, "subject")
		if subject == "" {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Missing subject", "subject path parameter is required"))
			return
		}

		// Verify the user exists before attempting an update.
		users, err := svc.List(r.Context())
		if err != nil {
			problem.Write(w, http.StatusInternalServerError,
				problem.New(problem.TypeInternal, r, "User list failed", err.Error()))
			return
		}
		found := false
		for _, u := range users {
			if u.Subject == subject {
				found = true
				break
			}
		}
		if !found {
			problem.Write(w, http.StatusNotFound,
				problem.New(problem.TypeNotFound, r, "User not found", subject))
			return
		}

		var body updateUserRequest
		if err := DecodeJSON(r, &body); err != nil {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Invalid request body", err.Error()))
			return
		}
		if body.Password == "" {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Missing password", "password is required"))
			return
		}
		if !validRole(body.Role) {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Invalid role",
					"role must be one of: admin, operator, viewer"))
			return
		}
		if err := svc.Put(r.Context(), subject, body.Password, body.Role); err != nil {
			problem.Write(w, http.StatusInternalServerError,
				problem.New(problem.TypeInternal, r, "User update failed", err.Error()))
			return
		}
		actor := identityFromCtx(r.Context())
		if rec != nil {
			rec.Record(audit.Entry{
				User:   actor,
				Action: audit.ActionUserUpdate,
				Note:   "subject=" + subject + " role=" + string(body.Role),
			})
		}
		w.WriteHeader(http.StatusNoContent)
	}
}

// DeleteUser handles DELETE /admin/users/{subject}. Returns 404 when
// the user does not exist and 409 when removing the subject would
// leave no admins.
func DeleteUser(svc UserAdminService, rec audit.Recorder) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		subject := chi.URLParam(r, "subject")
		if subject == "" {
			problem.Write(w, http.StatusBadRequest,
				problem.New(problem.TypeValidation, r, "Missing subject", "subject path parameter is required"))
			return
		}
		err := svc.Delete(r.Context(), subject)
		if errors.Is(err, sqlite.ErrUserNotFound) {
			problem.Write(w, http.StatusNotFound,
				problem.New(problem.TypeNotFound, r, "User not found", subject))
			return
		}
		if errors.Is(err, sqlite.ErrLastAdmin) {
			problem.Write(w, http.StatusConflict,
				problem.New(problem.TypeConflict, r, "Cannot remove last admin",
					"at least one admin user must remain"))
			return
		}
		if err != nil {
			problem.Write(w, http.StatusInternalServerError,
				problem.New(problem.TypeInternal, r, "User deletion failed", err.Error()))
			return
		}
		actor := identityFromCtx(r.Context())
		if rec != nil {
			rec.Record(audit.Entry{
				User:   actor,
				Action: audit.ActionUserDelete,
				Note:   "subject=" + subject,
			})
		}
		w.WriteHeader(http.StatusNoContent)
	}
}

// ListUsersV2 handles GET /admin/users. Returns every user sorted by
// subject. Password hashes are never included in the response.
func ListUsersV2(svc UserAdminService) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		rows, err := svc.List(r.Context())
		if err != nil {
			problem.Write(w, http.StatusInternalServerError,
				problem.New(problem.TypeInternal, r, "User list failed", err.Error()))
			return
		}
		out := make([]userListEntry, 0, len(rows))
		for _, row := range rows {
			out = append(out, userListEntry{
				Subject:    row.Subject,
				Role:       row.Role,
				CreatedAt:  row.CreatedAt,
				LastSeenAt: row.LastSeenAt,
			})
		}
		JSON(w, http.StatusOK, out)
	}
}